+++ proxy/contracts/predeployed/TokenFactory.sol
@@ -28,7 +28,7 @@ import "openzeppelin-solidity/contracts/token/ERC721/ERC721MetadataMintable.sol"
 
 contract ERC20OnChain is ERC20Detailed, ERC20Mintable {
 
-    uint private _totalSupplyOnMainnet;
+    uint256 private _totalSupplyOnMainnet;
 
     address private addressOfErc20Module;
 
@@ -46,11 +46,11 @@ contract ERC20OnChain is ERC20Detailed, ERC20Mintable {
         addressOfErc20Module = erc20Module;
     }
 
-    function totalSupplyOnMainnet() external view returns (uint) {
+    function totalSupplyOnMainnet() external view returns (uint256) {
         return _totalSupplyOnMainnet;
     }
 
-    function setTotalSupplyOnMainnet(uint newTotalSupply) external {
+    function setTotalSupplyOnMainnet(uint256 newTotalSupply) external {
         require(addressOfErc20Module == msg.sender, "Call does not go from ERC20Module");
         _totalSupplyOnMainnet = newTotalSupply;
     }
@@ -63,7 +63,7 @@ contract ERC20OnChain is ERC20Detailed, ERC20Mintable {
         _burnFrom(account, amount);
     }
 
-    function _mint(address account, uint value) internal {
+    function _mint(address account, uint256 value) internal {
         require(totalSupply().add(value) <= _totalSupplyOnMainnet, "Total supply on mainnet exceeded");
         super._mint(account, value);
     }
@@ -172,20 +172,20 @@ contract TokenFactory is PermissionsForSchain {
         assembly {
             nameLength := mload(add(data, 129))
         }
-        name = new string(uint(nameLength));
-        for (uint i = 0; i < uint(nameLength); i++) {
+        name = new string(uint256(nameLength));
+        for (uint256 i = 0; i < uint256(nameLength); i++) {
             bytes(name)[i] = data[129 + i];
         }
-        uint lengthOfName = uint(nameLength);
+        uint256 lengthOfName = uint256(nameLength);
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             symbolLength := mload(add(data, add(161, lengthOfName)))
         }
-        symbol = new string(uint(symbolLength));
-        for (uint i = 0; i < uint(symbolLength); i++) {
+        symbol = new string(uint256(symbolLength));
+        for (uint256 i = 0; i < uint256(symbolLength); i++) {
             bytes(symbol)[i] = data[161 + lengthOfName + i];
         }
-        uint lengthOfSymbol = uint(symbolLength);
+        uint256 lengthOfSymbol = uint256(symbolLength);
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             decimals := mload(add(data,
@@ -215,17 +215,17 @@ contract TokenFactory is PermissionsForSchain {
         assembly {
             nameLength := mload(add(data, 129))
         }
-        name = new string(uint(nameLength));
-        for (uint i = 0; i < uint(nameLength); i++) {
+        name = new string(uint256(nameLength));
+        for (uint256 i = 0; i < uint256(nameLength); i++) {
             bytes(name)[i] = data[129 + i];
         }
-        uint lengthOfName = uint(nameLength);
+        uint256 lengthOfName = uint256(nameLength);
         // solium-disable-next-line security/no-inline-assembly
         assembly {
             symbolLength := mload(add(data, add(161, lengthOfName)))
         }
-        symbol = new string(uint(symbolLength));
-        for (uint i = 0; i < uint(symbolLength); i++) {
+        symbol = new string(uint256(symbolLength));
+        for (uint256 i = 0; i < uint256(symbolLength); i++) {
             bytes(symbol)[i] = data[161 + lengthOfName + i];
         }
     }
